Fix global invocation size

+++ src/sim/GpuAlgorithm.hpp
@@ -62,7 +62,7 @@ class GpuAlgorithm
 
         // Note: The max warp size of nvidia cards is 32
         const uint32_t local_group_size = 32; // Attention: This value must match in the shader!
-        const uint32_t global_group_size = (uint32_t)ceil(Config::num_entities / local_group_size);
+        const uint32_t global_group_size = (uint32_t)ceil((float)(Config::num_entities) / (float)local_group_size);
         algo = _mgr->algorithm<float, Parameter>(tensors, _shader, {global_group_size, 1, 1}, {}, {parameter});
         shaderSeq = _mgr->sequence();
     }
